pyproject.toml
[project]
name = "aware-npc-chat"
version = "0.1.0"
description = "Chat with NPC aware of the world they live in "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "gradio>=5.32.0",
    "lxml>=5.4.0",
    "networkx>=3.5",
    "patool>=4.0.1",
    "pydantic>=2.11.5",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.12",
]
